Fix ArrayAccess declaration in the generated parser
PHP_ParserGenerator uses "ArrayAccess" without any namespace prefix for compatibility
with PHP 5.2. This poses a problem whenever the parser is defined as part of a namespace
because it then looks for ArrayAccess in that namespace.

Previously, we relied on a custom template to use "\ArrayAccess" instead.
This is hackish though, and an unnecessary burden to maintain.
Now, this is properly handled by PHP_ParserGenerator and the hack is not necessary
anymore.